refactor(BlackJackConstants): 상수를 오브젝트로 묶고 파일명 변경
