--- src/main/kotlin/domain/constant/BlackJackConstValues.kt
@@ -1,7 +1,0 @@
-package domain.constant
-
-const val BLACK_JACK = 21
-const val NOTHING = 0
-const val BIG_ACE = 11
-const val SMALL_ACE = 1
-const val DEALER_STAND_CONDITION = 16
