+++ src/main/kotlin/domain/constant/BlackJackConstants.kt
@@ -1,0 +1,9 @@
+package domain.constant
+
+object BlackJackConstants {
+    const val BLACK_JACK = 21
+    const val NOTHING = 0
+    const val BIG_ACE = 11
+    const val SMALL_ACE = 1
+    const val DEALER_STAND_CONDITION = 16
+}
